everything is a pointer
