+++ pkg/controller/admin/admin.go
@@ -29,7 +29,7 @@ type Controller struct {
 	config       *config.ServerConfig
 	db           *database.Database
 	authProvider auth.Provider
-	h            render.Renderer
+	h            *render.Renderer
 	limiter      limiter.Store
 }
 
@@ -39,7 +39,7 @@ func New(
 	db *database.Database,
 	authProvider auth.Provider,
 	limiter limiter.Store,
-	h render.Renderer,
+	h *render.Renderer,
 ) *Controller {
 	return &Controller{
 		config:       config,
