+++ pkg/controller/middleware/sessions.go
@@ -37,7 +37,7 @@ const (
 // request's context for future retrieval. It also ensures the flash data is
 // populated in the template map. Any handler that wants to utilize sessions
 // should use this middleware.
-func RequireSession(store sessions.Store, h render.Renderer) func(http.Handler) http.Handler {
+func RequireSession(store sessions.Store, h *render.Renderer) func(http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			ctx := r.Context()
@@ -101,7 +101,7 @@ func RequireSession(store sessions.Store, h render.Renderer) func(http.Handler) 
 
 // CheckSessionIdleNoAuth is an explicit check for session idleness. This check is also performed along with authentication
 // and is intended to be used when no other auth check is performed.
-func CheckSessionIdleNoAuth(h render.Renderer, sessionIdleTTL time.Duration) mux.MiddlewareFunc {
+func CheckSessionIdleNoAuth(h *render.Renderer, sessionIdleTTL time.Duration) mux.MiddlewareFunc {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			ctx := r.Context()
