+++ pkg/controller/modeler/modeler.go
@@ -36,12 +36,12 @@ import (
 type Controller struct {
 	config  *config.Modeler
 	db      *database.Database
-	h       render.Renderer
+	h       *render.Renderer
 	limiter limiter.Store
 }
 
 // New creates a new modeler controller.
-func New(ctx context.Context, config *config.Modeler, db *database.Database, limiter limiter.Store, h render.Renderer) *Controller {
+func New(ctx context.Context, config *config.Modeler, db *database.Database, limiter limiter.Store, h *render.Renderer) *Controller {
 	return &Controller{
 		config:  config,
 		db:      db,
